LayoutTests/platform/glib/http/tests/websocket/tests/hybi/handshake-fail-by-wrong-accept-header-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-wrong-accept-header' failed: Server returned incorrect “Sec-WebSocket-Accept” key
Test whether WebSocket handshake fails if the server sends incorrect Sec-WebSocket-Accept header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
